requirements.txt
python-dotenv
streamlit
langchain
langchain-groq
pdfplumber
faiss-gpu
ollama